[refactor] 토큰 생성 및 인증 함수 분리
- lib 폴더의 jwt-token 파일로 jwt토큰 생성 및 인증함수 이동
